cleaned up/speeded up is_unitary_test_big_SIMPLE_interrupt
The "up_mu_heights" function just runs into trouble on a few
examples (where there are many relevant larged K-types, but the
functions fails to find them. This did not introduce errors, but
failing to find to_ht targets meant that is_unitary was often
called unnecessarily. I think the present version mostly gives
[up] on up_mu; a fair amount of testing indicates that it's pretty
fast. (Jeff had trouble with three (x,lambda) pairs in E7_s, which
with an early up_mu took 1, 1, and 3 hour; on his previous fastest run
they took 5, 1.5, and 3 minutes. On my laptop now these three
now run in 21, 28, and 59 seconds.

There are a lot of flags, including some new ones, to play with, but I
think it's reasonable to try anything in E7 or E8 with the defaults
(except for more_shift, which should probably be 20-40?? The recent
code is much smarter when bottom_layer fails to help, so I hope that
the value of a large more_shift and of doing hash2 for small x are
both reduced.)