lots of speed (??) reorganizing
When x is too small for bottom_layer to work, the code used to
revert to hash2. Now instead it reverts to one_level. The original
behavior can be recovered with

one_level_revert_flag := false

(The change is that this used to be the default.)

There is a new global variable hash2_frac. When the KGB number is less
than (#KGB * hash2_frac), bottom layer reverts to hash_one_level (or
to hash2 if one_level_revert_flag is false).

I changed the defaults for mu_rank_mult to 1/2 and mu_time_mult to
1/1.

I changed (again) the way that sequences of test heights are computed,
making the sequence more widely spaced (with a spacing VERY approximately
proportional to the rank(G)*mu_rank_mult. I think this spacing is
good, but that the _number_ of test heights is probably now too
big. That is, the old code sometimes tested to a maximum height
H0. The new code uses the same number of more widely spaced heights,
so the max is some larger H1. I think that it would be best to reduce
the number of hts so that the max is closer to H0, but I haven't
tested to verify that. Clear as mud?