Fix the design of |type::wrap_tuple|: caller should pass the |floor| value.
The error was to depend on the component types passed to determine the |floor|
value as the minimal one needed; this overlooks the fact that the |threshold|
value of a type can be used for other purposes than to determine the status of
type variables used in the stored |type_expr|. Notably in the case of an empty
tuple type, a value |0| was always stored, for lack of any components to take
the |floor()| value from. The threshold of each component type must however be
used while packing its |type_expr|, which omission was also fixed.