Correct and simplify type checking of loops: |while|, |for|, and counted-|for|.
There were bugs due to recent clerical errors (forgotting to take a component
type and to export the body type of a loop to the loop itself). We took the
occasion to systematise the approach of different types of loops, making them
more similar, and on several occasions sharing identical pieces of code.