instrumenting FPP
This commit makes a number of small changes, mostly aimed at making it
possible to see the effectiveness of various parameters in the unitary
dual algorithms. A central feature is a new function

test_bl@(KGBElt, ratvec) = bool

which is meant to examine a pair (x,lambda), decide whether the bottom
layer argument is likely to work better than one_level for this
(x,lambda), and (if yes) pick an appropriate value of
more_shift_level.

The function is defined around line 89 of FPP_globalDirac.at. Right
now (because test_bl_flag defaults to "false" the function does
nothing at all: bottom layer is always called if the user asks for it.

I'll write some notes about how to investigate.