Fix some naming and multiple definition issues in the scripts
A problem came to light that the (constant) variable |one| was defined both in
bigMatrices.at and in lazy_lists.at (causing transient errors dependent on the
order of loading scripts), which was resolved by renaming both. There was one
more instance of Redefining, in sort.at, which was a simple copy-paste error.