Lots of speed improvements...
...mostly just corrections of typos, or adding something very
easy. For example, in one_param_test_bottom@(Param p), the first step
was to finalize p to a sum of various q, and to look up whether all
the q's were known to be unitary. The typo was that I looked up not q but
the (non-final) p; so this helpful test always failed. (The example
where Jeff found if had q unipotent, so actually running the
is_unitary test on q took a very long time.)

Anyway these changes seem to fix all of the ten worst times in Jeff's
E7_s run. We'll see what the next problem is!