Sorted non-minimal-Q bottom layer K-types, increasing by height.
I think that the bottom layer testing proceeds upward through this
list, stopping if it finds a non-unitarity proof. That ought to help
speed a little bit, but it's very difficult to test meaningfully.

The remaining to-do is for each bottom layer K-type, find the SMALLEST
Q for which it's bottom layer, and do the testing there. (Now the
testing is done on some maximal Q; so a bigger L, and more difficult
calculations on L.)