Modify |type::unify_specialize| to treat easy cases early, as now does |unify|
The easy cases are those where one or the other type or pattern being unified
can be assigned the other directly; when this is the case one should avoid doing
any expansion on the other, which would show up the the user in the form of
defined type names getting expanded to there definition where there seems no
reason to do so. Concretely, casting to a defined type now has that as its type.