Correct error in else_dumb...
...which caused slow execution but no wrong answers. The list of
candidate facets was constructed first using FULL K-characters rather
than the intended to_ht(p, hts[0]).

At the same time, noticed that low@Param was defined (differently) in
two different places. Now the desired definition is in the earliest
location in kl.at, and the use of that in deciding whether to do to_ht
is...well, made as originally intended; I'm sure it's far from
optimal.

The decision about whether to calculate a character just to height, or
to calculate the full character and truncate, now depends on a rat
trunc_factor, defaulting to 2/1. Increasing it does LESS truncation,
decreasing does MORE.