Remove some error types from error.h, and pass on |std::bad_alloc| unchanged
The latter concerns the KL code, which used to convert a |std::bad_alloc| into
an |error::MemoryOverflow| (now removed) which was not caught by the atlas
interpreter (which does catch type derived from |std::exception|, and therefore
would have caught |std::bad_alloc|). Not catching this kind of error is
undesirable, do this was fixed by changing the rebranding |throw| statement
into a simple |throw;| to pass on the error unchanged. In the mean time some
other error types like |error:FatalError| were removed (and the code using it
adapted); we did leave |error::NumericOverflow| which is thrown and then caught
inside the library, so that it is less prone to inadvertent dropping of errors.