Use K_chars_to_ht in one_level_end
This is what's used to finish bottom_layer argument, so it should save
a little time and memory there. As usual it's difficult to test that
in small G, so I'll commit and do some testing on a big computer.

Should do the same with hash2_end, which is invoked by
one_level_revert_flag:=false.