simplified lots...
...possibly making some things a little slower??  Realized that I
didn't need to keep making new pol_hash: one is plenty.

next_heights formerly made this sequence of heights grow a little bit
rapidly. Now that is mostly disabled, so that the sequence is a little
smaller. That is definitely a good thing close to x_open; it may slow
the code slighly for smaller x???

The script bottom_layer.at decides how far to look in bottom_layer on
the minimal Q using a function next_heights(RealForm)=int (applied to the
Levi of Q). This value was too small on E7_s, leading to bottom layer
calculations in E8_s that were not as good as they could have
been. The function is more or less proportional to a (new) global rat
NH_factor, which defaults to 1/3. Using

NH_factor:=2/3

seems to improve matters for large x in E8_s (???).

TO DO: the function test_bl in FPP_globalDirac.at decides whether to
use bottom_layer when it's called for, and what more_shift should be.
This function makes a lot of mistakes that are quite costly in E8_s. To
do is fixing the function. For now,

test_bl_flag := false

disables the function, so the user gets bottom_layer if it's asked
for.

bottom_layer also has an internal decision about whether to revert to
another algorithm, and this also is flawed. (Sometimes reversion is
good, but not in every case that the code does it.) This internal
reversion can be disabled with

revert_flag:=false