Correct errors related to red_count_flag TRUE
General cleaning up for hash2 and one_level, hoping to restore
something like the speeds of spherical unitary that we used to get.

There was serious error in a version of localFD_KHash_01_graph taking
a HT argument, leading to job aborts when red_count_flag = true and
quick_flag = false. That's fixed. The speeds are OK in E7_s spherical;
I'll test E8_s.

[[FaceVertsIndex]] LF, [[int]] eq, [vec] classListByFace)