speed improvements for very large vertex sets
For reasons I don't understand well, the function local_testK_hash
(defined in FPP_localDirac.at), which discards vertices and edges
failing is_unitary to a specified height, is sometimes slow with a
large vertex set and a very small test height; but faster with the
same vertex set and a larger test height. I have therefore changed
the default starting test height from number 0 on the list to
edge_offset, where edge_offset is a global variable defaulting to 2.