speed tweaks...
Most serious is restoring Jeff's idea of a global variable
KGB_frac. When x.number is smaller than (#KGB)*KGB_frac, bottom_layer
reverts to one_level. Default value is 2/3, which is probably too
large.

After bottom_layer (which looks only at vertices and edges) the
completion functions one_level_end and hash2_end now begin with
is_unitary_to_ht for the vertices and edges. The result is that
bottom_layer cannot be much worse than one_level (or hash2), because
the bottom_layer computations are almost always much faster than the
to_ht ones: if they are useless, then it's a small amount of lost
time, and if they succeed, everything else is faster.

Recently added a variable NH_factor, which used to default to
1/3. Increasing it means checking MORE bottom_layer K-types. The new
default is 1/2, which is a large improvement for some of Jeff's slow
E7_s cases (like (KGB element #19200,[  2,  1,  1,  2, -1,  1,  2 ]/1).