Change handling of |let () = E1 in E2| to provide a strong void context for E1
Due to the fact that we no longer set the expected type for E1 to the "pattern
type" for the left hand pattern before type checking E1 (a test for
compatibility is instead performed after finding a type), code that provided an
expression E1 with non void type and that used to pass began producing errors.
The same would also be the case for expressions that need a void context in
order to compile at all, even though that is rare and was not observed in the
library. Type checking of |let| expressions now makes an exception to the rule
that the RHS of the definition is type checked without any imposed type, by
imposing a void type when the LHS is "()". This does not revert exactly to the
old situation but is good enough in practice, and can be quite easily achieved.