cleaning up...
When directed graphs of parameters are traversed, usually one starts
at the LAST vertex, because from there (expected) non-unitarity
results have the most consequences.  Started a project of attacking
the graphs from both ends: eight nodes from the end, then one from the
beginning... This is only in the functions local_testK_hash and
local_testK_hash_know_lower so far. It's awkward, but this version
appears to run correctly and not to increase time noticeably. (In
spherical E7_s, it DOES increase memory use by about 10%.

I will keep working on this to try to make it faster and smaller, but
at the moment I believe it is reliable, and MUCH faster than anything
from May 18 2025 or earlier.