More speed tweaks...
There were one or two mistakes that (rarely) took a lot of time
without being wrong: passed is_unitary_test_big_SIMPLE_interrupt an
integer argument that I thought was a first height to test to; but in
fact it was an INDEX into the list of next heights. So in those cases
the code did a to_height calculation to the 100th height in some list
rather than to height 100. I think this explained some of Jeff's bad
(x,lambda).