Fix bug in applied identifier resolution: replace |logic_error| by |type_error|
The bug was that if the context of an applied identifier requires a function
type without any restrictions on its argument type, as when processing |g| in
|compose(f)(g)|, then it can still impose a restricting on the result type, and
should therefore be ready to produce a |type_error| even when exactly one
definition of |g| is present. We had thrown a |logic_error| instead.