Redo handling of identifier from overload table when used without operator cast
The rule for using an identifier only known in the overload table becomes that
an operator cast can be omitted provided that a single defined variant matches
the imposed context type. If the context wants a non-function type, this cannot
possibly succeed, and so no attempt is even made to use the overload table. If
to the contrary it requires an explicit function type, both argument and result
part of that type are used in selecting the proper overloaded instance.