fixed errors in implementation of saved nonunitary to ht
One was failure to add to the known nonunitary height list when adding
to the known non-unitary, in one case of about twenty.

Second was neglecting to consider that there might _be_ no stored
nonunitary height (which as recorded as an entry of -1 in the list).

I'm in desperate need of advice on how to rearrange such code (hash
tables with associated lists of data for each entry) so that it's
easier to maintain!