changed a couple of defaults for speed...
...KNU_flag (which stands for "known nonunitary") now defaults to
true, meaning that the functions examining a set of faces for
unitarity to ht now check big_unitary_hash for faces already known to
be nonunitary to that height. This costs a few percent in time for
E6_s and E6_q, but helps (at least 2% up to KGB=15000) in E7_s

edge_offset and edge_skip now default to 1 (previously 2). This helps
a LOT in E6_s and E6_q, and seems to help in E7_s.