Many minor changes to bring endoscopy into compliance: check_scripts.sh should succeed, and there should not be any extraneous files in the repository
This is preliminary to merging master, which should be straightforward. This is preliminary to merging with
master-2.0, which is going to be quite involved.

Most of the changes involved adding some file loads at the top of files. Some commands, which had been defined in other
files, were moved to basic.at.