tweaks for speed...
...changing the defaults for edge_skip and edge_offset to 1 and 1 was
to avoid difficult is_unitary_to_ht calculations. Turns out that a
better way to do that is to use is_unitary_to_hts, testing small hts
first and only the larger ones if that fails.

Cleaned up the output of local_test_hash a little. Altogether no
serious changes, but a lot of gain in speed.