reorder some tests for speed improvement
When the code is running through a a series of is_unitary (or to_ht)
tests, the precise order in which the tests are done affects the speed
(because both the scripts and the library record some of what they
learn with each test).

Main change here is in one_level, on the last is_unitary pass: the
script now reverses the order of each list LFDKH[d] of d-dimensional
faces. For reasons I do not understand, this reversal reduces my
laptop time to compute the spherical unitary dual of E7_s from 400
seconds to 330 seconds.

This could be an accident about E7_s; I will do some testing to see
whether there are parallel savings for other G. If so, I will (by
Friday 6 June?) make parallel changes to hash2.