Rewrite hash.at to use a generic hash table type and constructor
Types that used to be called something like |RootDatum_hash| are now instances
of a single type constructor, as |Hash<RootDatum>|, while the way to create an
instance is still to call |make_RootDatum_hash()|, so not much changes for
clients. The old type names are still defined as aliases to facilite transition
to the new version.