Replace some instances of implicit conversions of kind M[V]
The |[vec]| to |mat| implicit conversion is dangerous since it will give en
error in case the list is empty (the number of rows of the empty matrix to be
returned cannot be deduced). The better alternative is to use the built-in
function |#@(int,[vec])| of which the first argument gives the common size of
the vectors, to become the number of rows of the matrix. While one use was
explicitly commented as being safe (the list of vectors cannot be empty), even
here it helps readability to explicitly specify the expected number of rows.