Change scanner to allow multi-character operators like ==, <<, <==, <==>
All combinations of <, > and = are allowed, and these operators always get
priority eual to 2. All except single '<' and single '>' can serve as a point to
break a line after, without the parser thinking a command ends there.

This change implies one need space in type expressions like Matrix<Maybe<int> >
(where both Matrix and Maybe are type constructors), since ">>" will now be
scanned as a single operator token, giving an errro. Such type expressions are
rare, but there was one instance in the library that needed to be modified.