Correct flaw in built-in |adjoint|, and iron out |simply_connected| as well
The flaw was that |adjoint@(LieType,bool)| did not work in the presence of a
nonzero dimensional central torus, due to passing a too small square matrix to
|root_datatum@(LieType,mat,bool)|. This probably was the case ever since the
|Cartan_matrix@LieType| built-in was modified to ignore any central torus and
return the true semisimple rank sized Cartan matrix. The internal method
|LieType::transpose_Cartan_matrix| does give a rank sized matrix, so we now use
that (with diagonal entries made 1 to get a full rank sublattice) instead of
invoking |Cartan_matrix@LieType|. All in, the only wrapper function that is
still called is that of |Cartan_matrix@LieType|. The built-in function
|simply_connected@(LieType,bool)| had no error, but was modified as well to make
the similarity and differences between these two built-in functions clearer.