made red_count work again
A flag red_count_flag is supposed to trigger using easy irreducibility
criteria to tell when two reps have the same character, and so
shorten character computations (of all edges and vertices at
(x,lambda), for example). The flag defaults to false.

For some time (?) "true" has been broken, because of a couple of dumb
coding mistakes in the function localFD_KHash_01_graph. These are now
fixed. But making the flag true still seems to hurt speed. I will
push this and test it on E8_s, where the more difficult character
computations might make it worthwhile.