Correct language implementation: function definitions and applied identifiers
A function definition was entered as variable if the type of the defining
expression was a user defined type name for a function type.

A name not known as a global identifier but present in the overload table was
accepted even if its usage was ambiguous, because the type recorded on the first
match was not properly cleared before attempting a second match. Also the error
message was polluted by using and not clearing a |std::ostringstream| variable
int the matching process, and then reusing the same variable.