Supplement to: Fix long-standing error in DiracBestTest
On the current repair branch the error was fixed directly after it was
introduced. Nonetheless we include this bugfix commit, since it make a few other
changes, so that we become identical to the original branch after the fix.

The original commit message said: the function definition

   set DiracBestTest(Param p, int HT) = bool:
       DiracBestTest(Dirac_best_steps, p,
       K_type_pol(character_formula_to_height(p,HT)))

introduced February 15 2025 in commit 811e3004a518d443, is wrong:
should be

   set DiracBestTest(Param p, int HT) = bool:
       DiracBestTest(Dirac_best_steps, p, HT,
       K_type_pol(character_formula_to_height(p,HT)))

The error means that while testing is_unitary_to_ht(*,HT), an
individual representation may (incorrectly) be found to fail Dirac
inequality, and so to to be incorrectly removed from the candidate
list. This happens in E7_s for at least two unitary reps, including

   parameter(E7_s, [0,1,1,0,1,1,1]/1,[-1,0,2,-4,2,2,2]/2)

at height 261. In principal it is also possible that a representation
in an equivalence class could be correctly identified as nonunitary,
but using a K-type above HT. (I never saw this happen.) in that case
the entire equivalenve would be removed from the candidate list,
although that would not be justified.

I believe the error can ONLY lead to omission of unitary
repreaentations, not to labelling nonunitary as unitary. But all
calculations from February 15, 2025 need to be repeated.
Steve at least has often used flags which print an output line like

   ...Dirac_best elims = 14278

In runs with several KGB or (x,lambda) pairs, this number is
cumulative over the run. If the number does not change from one
(x,lambda) to the next, than the error has no chance to arise, snd the
unitary output produced in that step is complete.