Restore |LinearSolution| type definition, allow |GL(0)| to be a trivial group
The first point serves to make |linear_solve|, which returns the union type now
names |LinearSolution| in basic.at, more comprehensible, because it attaches
tags to the union variants anyway. These tags are now |empty_set| and |solution|
(the latter used to be |affine_subspace| but that name actually conflicts with
a type definition present in Gaussian_elim.at, so we changed it).

Making the call |GL_roots(0)| valid, and thereby |GL(0)|, allows obtaining all
permutations of |n| effectively as |W(GL(n))|, with no problems for |n=0|.

While revising groups.at, the calls of |root_datum| with implicit value |false|
of |prefers_coroots| were replaced by calls with an explicit value. This value
was not always |false|, but set more or less along the convention that the value
is |false| for |adjoint| root data and |true| for |simply_connected| ones; this
obviously cannot be a hard rule, and when it is unclear we still set it |false|.