Add a new script diagram for diagram automorphisms, and supporting changes
One supporting change is the introduction of an |Iterator| type constructor, to
model values that can produce a stream of values by being called repeatedly,
without having to store the entire stream of values all at once. This is used in
diagram.at to generate the permutations of the diagram dynamically.