Extend diagram automorphisms to those of root data.
This does not pretend to find all automorphisms, only those that on the central
torus part are plus or minus the identity (and at most one per diagram
automorphism, so if plus identity works, its negative is not tried). As long as
the central torus has rank at most 1, it does appear to find all these
automorphisms. This is again done via the definition of an iterator value (to
which |to_list| can be applied if useful), but one can do without that if one
just wants to traverse the automorphisms, as illustrated in |list_automorphisms|.

This has a limited usefulness of finding the outer automorphisms of a (pseudo)
Levi subgroups, by listing the automorphisms of their root datum: not all such
automorphism will be found in the common case that the co-rank of the Levis is
2 or more. For instance it will not find the extension of the A1A1 automorphism
for |Levi_datum (GL(4),[0,2])|, although it will for |Levi_datum (SL(4),[0,2])|.