Rewrite script conjugate.at using more efficient methods, and adapt other scripts
The main goal of this change is to remove constructing new |RootDatum| values
all the time, by working systematically in the original root datum and
representing sub-data for Levi subgroups by |LeviInfo|. This requires some
functions to change their type.

Parts of conjugate.at were removed, and clients were made to use the new
functionality that replaced it.

An obvious bug in |normalizer@(RootDatum,RootDatum,[mat])| was removed: the
(outer automorphism) permutation matrix |P| of size the semisimple rank can only
be intended to permute the simple roots, not their coordinates, so it should be
applied as |simple_roots(rd)*P| rather than (as it was) |P*simple_roots(rd)|.