Complete revision of conjugate.at, removing computation with |RootDatum| values
The version now favors using |(RootDatum,[int])| to represent (pseudo) Levi
subsystems by a list of their simple roots as indices into the root datum.

We believe having removed a serious misunderstanding concerning the functions
|outer_automorphism_group| and |normalizer|. The former seems to be in bijection
with the set of automorphisms of the diagram of the subsystem that can be
realized by an element of the ambient Weyl group $W$, which more properly is the
quotient of the normalizer in $W$ of the subsystem by its centralizer. Moreover
the use of |outer_automorphism_group| was unclear and inconsistent (it was being
computed for the derived group of the Levi subgroup, making the action matrices
essentially useless), and |normalizer| made no attempt to incorporate the
centralizer in $W$ of the subsystem. Also |normalizer_report| in |subgroups.at|
was somehow dividing Out/Norm, which is not something I could make any sense of.

So these functions and their clients were rewritten.