corrected an old error when red_count_flag:=true (default false)
The two functions localFD_KHashbt_lookDownGEO in FPP_faces_herm.at are
meant to add higher dimensional faces, including the coordinate
identifying their K-characters. When red_count_flag is true, it's
sometimes possible to conclude that the K-character is one already
calculated on a subface, so no need to recompute. When this happened,
I miscalculated the location of the coordinate with the known
K-character.

Even when red_count_flag was true, this often caused no trouble; the
graphs calculated were wrong, but at this stage of the calculation the
graphs had a lot of isolated points.

I eventually figured that there was a problem with the K-character
index, so added a function check_polys that checks that. If
check_polys_flag is assigned to true (default is false),  this test is
run often in the one_level calculation. For now I don't see any need
to use the test, but it's conceivable that I'll break something in the
future, so I've left it. DID NOT add it to the other algorithms; that
would be very reasonable to do.

In E7_s spherical, red_count_flag:=true reduces the time to calculate
initial K-characters from 100 seconds to 75 seconds on my laptop; but
calculating the red_count's takes 115 seconds, so there is a net loss
of 90 seconds. But this approach is not hopeless...