File by_dim.at checks unitarity of candidate list.
Starts with two Param_hash's cand_hash and conf_hash of
representations of G. Reps in conf_hash are known to be unitary, and
reps in cand_hash are hoped to be unitary. Each family is divided into
[Param_hash] cand_hashes and [Param_hash] cand_hashes, indexed by
dimension of infinitesimal character face. Starting with largest
dimension, function

deform_test_all(G,cand_hash,conf_hash)

tests unitarity of each p in cand_hashes[d] that is NOT in
conf_hashes[d]. If test fails, the function stops with a warning. If
it passes, then p is deformed to all e-dimensional boundary faces of
its face, and those deformed reps are added to conf_hashes[e].

If some deformed rep is not present in cand_hashes[e], a warning is
printed, but the function continues.

At the end conf_hashes contains all the reps in cand_hash and their
deformations (unless some rep in cand_hash was not unitary).

At the moment this works only for equal rank simple G.