modified by_dim.at...
Tests showed that the E8 confirmation slowed on some 2-diml faces for
E8_s, where is_unitary took a very long time; but unitarity was the
same as on a 1-diml edge where the computation was much faster.

When look_down_flag is false (default), the code is unchanged. When
look_down_flag:=true, the code looks for such ways to speed
is_unitary. Small testing suggests new code might be correct; not yet
tested for speed.