Corrected serious error...
...the function is_unitary_to_hts_big_SIMPLE_interrupt, when it found a
rep to be nonunitary, was incorrectly recording the height to which
that discovery was made. As a result some nonunitary reps that were
actually unitary to a certain height were recorded as nonunitary to
that height. Then they could appear in an equivalence class to height,
and cause that class to be incorrectly labeled as nonunitary (when in
fact only SOME of its members were nonunitary).

This took several days to find. I think there is still an error
related to red_count (only appearing when red_count_flag is changed to
true). But I won't look for it soon.