Relax identification of tabled types, so equal ones can be printed differently
When on defines for instance |set_type Partition = [int]|, we used to identify
every |[int]| installed in the internal |type_map| (namely, occurring anywhere
in a multiple |set_type|, or as full right hand side of a simple |set_type|)
with |Partition|, notably using that identifier to represent it on output. This
is confusing to users, so we now allow different representations of structurally
equivalent types to coexist in |type_map|, so that the output of a type can more
faithfully correspond to the way the type was originally specified. This does
suggest that name equivalence of type is used internally when actually it is
structural equivalence, potentially causing confusion; one cannot have it all.