Add a built-ins |Weyl_orbit_ws| and |Weyl_orbit| for Weyl subgroup orbits
These both come in two forms, one for weights and one for coweights (where as
usual the elements from |Weyl_orbit_ws| are taken to be right-multiplied in the
case of coweights to get the orbit). The |Weyl_orbit| forms could easily be
user-defined in terms of the |Weyl_orbit_ws| forms, but internally the
implementation does not depend on applying general (non reflection) Weyl group
elements to weights or coweights, so the built-in forms are faster than that.

In all cases the subgroup is specified by root indices for simple generators,
this can be (positive or negative) any list defining a valid Cartan matrix.