correct FPP_unitary_hash_steve
This version seems to give good answers for a bunch of groups up to
real rank 5. Still slower than one_level by factor of several; perhaps
a bit less memory?? Anyway I want to commit this before I find more
bugs.

The problem was move(WeylElt y), which carries (x,lambda) to
(x1,lambda1) defining the same LKT. x1 = cross(y,x) was fine, but the
formula for lambda1 is a pain. It's now in move_theta in steve.at; I'm
optimistic, but not yet certain.

Next plan is to remove a bunch of "asserts" which were helpful for
finding bugs, but now just slow things down.