Add |append| and alias of |#@T([[T]],[T])| in basic.at for user convenience
The convenience lies in the fact that this uncouples from the "prepend" meaning
of |#| (i.e., form |#@T([T],[[T]])|) so that |append(LL,[])| is unambiguous
for |LL| of some type |[[T]]|, whereas |LL # [] | is ambiguous. As a consequence
one can write |x append:= []| when |x| is a variable of some list-of-list type,
or even |a[i] append:= []| when |a| is a variable of list-of-list-of-list type.