Minor C++ changes: spelling "too big" and argument type of |basic_io::put_int|
The latter was sent to |std::uint32_t| to make clear that it only outputs the
least significant 4 bytes. The change was prompted by warning messages from the
filekl compilation unit, where a |BlockElt| constant near the largest value for
its (|long unsigned int|) range was passed to |put_int|. The fact that this is
deemed OK in this instance is now made explicit by using a |static_cast|.

In passing we changed the Makefile so that the 4 files dependent on filekl.w our
now built automatically when running 'make', and these dependent files were
therefore removed as items separate from filekl.w in the git repository.