Add a new version of |assert| that can be dynamically disabled for speed
Since expensive |assert| calls are so because their Boolean value takes long to
compute, any speed-up must avoid this evaluation, which is why the new |assert|
instances take a (->bool) rather than bool argument. The old definitions are
still given at the end of basic.at, but none of the basic function use them.
Therefore old scripts will continue to compile, but their |assert| calls should
be updated to make disabling possible; this is as easy as placing |@:| before
the condition to be tested (so that |assert| might decide not to evaluate it),
in the two-argument case optionally followed by parentheses grouping those two
arguments (so that even the pair-forming of the arguments can be disabled).

The dynamic switch is a simple boolean variable called |seat_belt_on|.