Fix an error in testing equality of defined types: name equality is insufficient
A recent change allowed us to distinguish type expressions obtained using
different type constructors, even if their expansions are equal. That change was
not correctly implemented, in that it forgot in the case of equality of type
(constructor) names to test whether they actually referred to equivalent types.
After this commit, equality of type names is a necessary, but not a sufficient
condition for new user defined types to be identified with existing ones.