Improve |sort_by| to make it as fast as is used to be
In generalizing |sort_by| to work with an attribute |f| of any type rather than
just |int| or |rat|, I did not consider that by wrapping |f| into the comparison
function, it gets called much more often than before. While keeping the
generality, we revert to the approach of first building a list of attributes (by
mapping |f| over the list to be sorted) and computing the |ranking| permutation
that allows sorting it; this same permutation then also sorts the original list.