Modify a number of scripts to get closer to their state on 'master' branch
We made few attempts to preserve any distinctive features not present on
'master' but we sometimes did not include extra stuff that is present there; the
main objective is that future changes here give few conflicts when then merging
to 'master'. On a few occasions we did make some improvements though.