Relax scanning rules for grouped type definitions: allow unapplied constructors
When scanning a grouped type definition, we do not yet know the set of type
names that it will introduce, and hence cannot know if a type constructor name
will be overridden in the current definition: if it is not, it should take its
usual number of arguments, but if it is not it should take none (as any type
recurrence implicitly uses an unchanged list of type arguments). Not allowing
for the latter scenario, as we did up until now, makes it difficult to reload a
file containing a recursive type constructor definition, as the recursive uses
of the defined types sill now be refused as being type constructors used without
argument list. Thus explicit |forget| commands would be required.

Allowing unapplied type constructors solves this at the price of a doubling of
the part of the grammar devoted to type expressions. Also the handling of
grouped type definitions now has the added responsibility of checking that all
uses have the correct number of type arguments; this is not yet done here.