fixed ind_cand_params...
and added ind_test@Param = int, which returns 1 if p is a Langlands quotient
of unitarily induced from a proper real parabolic, -1 if it is unitarily
induced from non-unitary, and 0 if neither.

The function is_u_ind@Param = bool just tests whether p is LQ of
unitarily induced from unitary.