Fix a change to |frequencies| and |repeat_parts| that was recently made
Since |frequencies| is intended to count parts in compositions of a natural
number so that |repeat_parts| can generate the associated (weakly decreasing)
partition from it, and parts equal to 0 are not represented in this partition,
the recent change was to no longer count those parts at all in |frequencies|, so
that |repeat_parts| did not even need to bother about not generating such parts.

However there were, on the 'master' branch, uses of |frequencies| outside of
combinatorics.at, where counting parts equal to 0 was depended upon; notably in
|diagram_D_to_partition| in springer_table_D.at. In order that |frequencies|
still do what it was expected to do there, we restored its functionality before
the change, and adapted |repeat_parts| to again ignore the count for zero parts.