added deformation awareness to FPP_unitary_hash_steve()
New functionality is activated by use_deform_steve(). When this is
done, every test in FPP_unitary_hash_steve (which begins by computing
a vec loc describing the location of the infinitesimal character with
respect to reducibility planes) considers also deformations loc_def of
loc. If one has already been found nonunitary, then loc is known to be
nonunitary. If one has been found unitary, and the K-character has not
changed, then loc is also known to be unitary. In the case of F4_s,
the former happens 55 times and the latter 12 times, leading to a time
saving of about 2.5%. Testing bigger G.