Change grammar to allow postfix-shriek inside formulas: | n! \ (k! * (n-k)!) |
For some reason the rule for formulas includes the alternatives for |primary|
(taken at a time before | n! | was introduced) in separate rules, rather than a
single rule using the non-terminal |primary| itself. As a consequence, adding
the factorial alternative for |primary| worked in isolation, but not as a
subsexpression in a formula. Merging those two rules to one solves the problem.