issue with standardizing a real Levi in SU(3,3)...
made a kludge fix so SU(3,3) works, but there is still a problem.

Issue is the function simplifier_theta in steve.at, which is meant to
carry a non-standard real Levi to a standard one. The old function
often worked, but failed in SU(3,3). Made a kludge so that SU(3,3)
works, and probably lots of other G, but it isn't complete.