Fix flaw in preprocessing grouped type definitions: avoid using |swallow| method
Normally we call |global_id_table->swallow| to transform types from the form in
which the parser prepares them to internally usable form, replacing for the case
of user defined types and type constructors the stored identifier number by the
appropriate |type_map| index. In the right hand sides of grouped type
definitions, special care was taken for the type identifiers defined in that
group, so that recursive uses can be associated with future |type_map| entries,
but in the case of existing user define types or type constructors, the method
|global_id_table->swallow| was still called. That is an error, because it makes
no precautions when recursively processing the type arguments of a constructor,
crashing the program in case these contained recursive uses of types being
defined in the group. The solution is to simulate what |swallow| would do, while
correcting the behavior. This required adding a manipulator method to
|type_expr| to make an in-place change to the number stored in a |tabled| node.